add union light client to main test suite
